requirements.txt
# Core dependencies
pychm>=0.8.6
beautifulsoup4>=4.12.0
markdownify>=0.11.6
markdown>=3.4.3
click>=8.1.3
tqdm>=4.65.0
gitpython>=3.1.31

# Development dependencies
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0